+++ rust/dbn/src/error.rs
@@ -51,7 +51,7 @@ pub type Result<T> = std::result::Result<T, Error>;
 impl From<csv::Error> for Error {
     fn from(value: csv::Error) -> Self {
         match value.into_kind() {
-            csv::ErrorKind::Io(io) => Self::io(io, "While writing CSV"),
+            csv::ErrorKind::Io(io) => Self::io(io, "while writing CSV"),
             csv::ErrorKind::Utf8 { pos, err } => {
                 Self::Encode(format!("UTF-8 error {err:?}{}", Self::opt_pos(&pos)))
             }